+++ R/daily_stats.R
@@ -1,5 +1,3 @@
-daily_stats <- function(site, year, core) {
-
   #' Produce daily stats for a site
   #'
   #' Produces a data frame of daily stats.
@@ -8,7 +6,10 @@ daily_stats <- function(site, year, core) {
   #' @param year Year of deployment
   #' @param core Core data frame, produced by `stitch_site`
   #' @return Data frame with one row for each day of depolyment, and columns with a number of statistics
+  #' @keywords internal
 
+
+daily_stats <- function(site, year, core) {
 
 
   cols <- c('Date', 'Min_DO', 'Min_DO_Time', 'Prop_Under_6', 'Duration_Under_6',