+++ R/get_file_hashes.R
@@ -1,5 +1,3 @@
-get_file_hashes <- function(files) {
-
    #' Get md5 hashes for a vector of files
    #'
    #' Given one or more file paths and names, return md5 hashes for each. Returns
@@ -7,7 +5,10 @@ get_file_hashes <- function(files) {
    #'
    #' @param files A list of full file specifications
    #' @return A matching vector of md5 hashes
+   #' @keywords internal
 
+
+get_file_hashes <- function(files) {
 
 
    z <- rep(NA, length(files))