+++ R/longest_duration.R
@@ -1,5 +1,3 @@
-longest_duration <- function(x) {
-
    #' Give duration of the longest run of low DO for each day
    #'
    #' Finds the longest run of DO below a threshold for each day.
@@ -15,8 +13,11 @@ longest_duration <- function(x) {
    #' 2. `Date_Time` The date and time, in format 2024-05-30 16:20:00
    #' 3. `trigger` TRUE when DO is below the selected threshold (the name of this column varies)
    #' @return Length of the longest run of TRUEs in fractional hours
+   #' @keywords internal
    # Source: sensemaking app, buzz.stats
 
+
+longest_duration <- function(x) {
 
 
    names(x)[3] <- 'source_trigger'                                                           # the trigger column doesn't have a consistent name, so give it one